+++ cincoconfig/config.py
@@ -10,7 +10,7 @@ from typing import Union, Any, Iterator, Tuple, Callable, List
 from argparse import Namespace
 from itertools import chain
 from .abc import Field, BaseConfig, BaseSchema, SchemaField, AnyField, ValidationError
-from .fields import IncludeField, InstanceMethodField
+from .fields import IncludeField, InstanceMethodField, SecureField, VirtualField
 from .formats import FormatRegistry, TFormatFactory
 
 
@@ -503,17 +503,19 @@ class Config(BaseConfig):
 
         return self.loads(content, format)
 
-    def dumps(self, format: str, **kwargs) -> bytes:
+    def dumps(self, format: str, virtual: bool = False, secure_mask: str = None,
+              **kwargs) -> bytes:
         '''
         Serialize the configuration to a string with the specified format.
 
         :param format: output format
+        :param virtual: include virtual fields in the output
+        :param secure_mask: replace secure values, see :meth:`to_tree`
         :param kwargs: additional keyword arguments to pass to the formatter's ``__init__()``
         :returns: serialized configuration file content
         '''
-        # formatter = FormatRegistry.get(format, **kwargs)
         formatter = FormatRegistry.get(format, **kwargs)
-        return formatter.dumps(self, self.to_tree())
+        return formatter.dumps(self, self.to_tree(virtual=virtual, secure_mask=secure_mask))
 
     def _process_includes(self, schema: BaseSchema, tree: dict,
                           format_factory: TFormatFactory) -> dict:
@@ -634,10 +636,22 @@ class Config(BaseConfig):
                 value = self._data[key]
             yield key, value
 
-    def to_tree(self) -> dict:
+    def to_tree(self, virtual: bool = False, secure_mask: str = None) -> dict:
         '''
         Convert the configuration values to a tree.
 
+        The *secure_mask* parameter is an optional string that will repalce
+        :class:`~cincoconfig.fields.SecureField` values in the tree.
+
+        - ``None`` (default) - call the field's :meth:`~cincoconfig.fields.SecureField.to_basic`
+          method.
+        - ``len(secure_mask) == 1`` (single character) - replace every character with the
+          ``secure_mask`` character. ``value = secure_mask * len(value)``
+        - ``len(secure_mask) != 1`` (empty or multicharacter string) - replace the entire value
+          with the ``secure_mask``.
+
+        :param virtual: include virtual field values in the tree
+        :param secure_mask: mask secure values with a string
         :returns: the basic tree containing all set values
         '''
         tree = {}
@@ -645,10 +659,22 @@ class Config(BaseConfig):
         fields.update(self._fields)
 
         for key, field in fields.items():
+            is_virtual = virtual and isinstance(field, VirtualField)
+            if key not in self._data and not is_virtual:
+                continue
+
             if isinstance(field, BaseSchema):
-                tree[key] = self._data[key].to_tree()
-            elif key in self._data:
-                tree[key] = field.to_basic(self, field.__getval__(self))
+                value = self._data[key].to_tree(virtual=virtual, secure_mask=secure_mask)
+            elif isinstance(field, SecureField) and secure_mask is not None:
+                value = self._data[key]
+                if len(secure_mask) == 1:
+                    value = secure_mask * len(value)
+                else:
+                    value = secure_mask
+            else:
+                value = field.to_basic(self, field.__getval__(self))
+
+            tree[key] = value
 
         return tree